requirements.txt
aiohttp
Requests
selenium
flask
Pillow


